Add a timeout to execute JITed code
Add a 1 second alarm on the child process (the one that executes the
JITed code generated by the parent) in order to avoid halt in the
special condition where we have a push of a register containing the
very own address of the push instruction followed by a return that
by its turn creates a never ending loop. Or in cases where a jump
or branch instruction target its own address, which also creates
a never ending loop condition.

Note that there is no need to set a signal handler on the child
since on getting the alarm clock signal the process is killed by
the kernel and the exit status is set ok to be read by the parent
that decides based upon it if it's or not a valid instruction.

Fix: https://github.com/gromero/panoptica/issues/1