Add detection for load/store or instruction fusion
Add detection for a load/store that access a not mapped VMA
address or a instruction fusion that fuses with return byte
and so removes the return condition from the code cache leading
to the execution of non-JITed code that yields a segfault.

Also improve comment on the self push/ret case and self branch
informing that the JITed actually halted (main symptom).