Isolate jiter from instruction generator
Now the jiter instruction accept a void pointer instead of a
int since it's a more flexible way to pass the generated code
to be executed to the jit(), which adds proper terminators for
the code.

Provide some debug control (-DDEBUG and dbgl) to see what the jit
code comes to be in the codecache.